Separate view classes by models
