+++ django_material_demo/cms/polls/views/file.py
@@ -1,0 +1,13 @@
+from material.frontend.views import ModelViewSet
+from polls.models import File
+
+
+class FileViewSet(ModelViewSet):
+    model = File
+    list_display = ['file_name', 'file_type', 'file_size', 'storage_loc']
+
+    def has_add_permission(self, request, obj=None):
+        return False
+
+    def has_change_permission(self, request, obj=None):
+        return False
