+++ django_material_demo/cms/polls/views/question.py
@@ -1,182 +1,16 @@
 from django import forms
-from django.contrib.auth.forms import UserChangeForm
 from django.core.exceptions import NON_FIELD_ERRORS, ValidationError
-from django.db.models import Count
-from django.forms import (BaseInlineFormSet, BooleanField, EmailField,
-                          ModelForm, model_to_dict)
-from django.forms.widgets import CheckboxInput, RadioSelect
-from django.http import HttpResponse
-from django.views import generic
-from django_filters import CharFilter, NumberFilter
-from library.django_superform import (ForeignKeyFormField, InlineFormSetField,
-                                      SuperModelForm)
+from django.forms import BaseInlineFormSet, BooleanField, ModelForm
+from django.forms.widgets import CheckboxInput
+from django_filters import CharFilter
+from library.django_superform import InlineFormSetField, SuperModelForm
 from material import Fieldset, Layout, Row
-from material.frontend.views import (CreateModelView, DetailModelView,
-                                     ListModelView, ModelViewSet,
-                                     UpdateModelView)
+from material.frontend.views import (CreateModelView, ListModelView,
+                                     ModelViewSet, UpdateModelView)
+from polls.models import Attachment, Choice, Question, QuestionFollower
 
-from polls.models import (Attachment, Choice, File, Question, QuestionFollower,
-                          User, Vote)
-
-from ..utils import (FieldDataMixin, FormSetForm, GetParamAsFormDataMixin,
-                     ListFilterView, NestedModelFormField, SearchAndFilterSet,
-                     get_html_list)
-
-
-class FileViewSet(ModelViewSet):
-    model = File
-    list_display = ['file_name', 'file_type', 'file_size', 'storage_loc']
-
-    def has_add_permission(self, request, obj=None):
-        return False
-
-    def has_change_permission(self, request, obj=None):
-        return False
-
-
-class AccountForm(UserChangeForm):
-    email = EmailField(required=True)
-
-    layout = Layout(
-        'username',
-        'email',
-        'password',
-        Row('first_name', 'last_name'),
-        'groups',
-        'user_permissions',
-        'is_active',
-        Row('is_staff', 'is_superuser'),
-        Row('date_joined', 'last_login'),
-    )
-
-
-class UserForm(SuperModelForm):
-    account = ForeignKeyFormField(AccountForm)
-
-    layout = Layout(
-        'account',
-        'group',
-        Row('subs_start', 'subs_expire'),
-        'followed_users'
-    )
-
-    form_widgets = {'group': RadioSelect}
-
-    class Meta:
-        model = User
-        fields = ['group', 'subs_start', 'subs_expire',
-                  'followed_users']
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-
-        if self.instance:
-            self.initial = model_to_dict(self.instance)
-
-            account_queryset = self.instance.account
-            self.initial["account"] = account_queryset
-
-
-class UserDetailView(DetailModelView):
-    def get_object_data(self):
-        user = super().get_object()
-
-        # Account fields
-        account = user.account
-
-        account_fields = [
-            'username', 'email',  # 'password',
-            'first_name', 'last_name',
-            'is_active', 'is_staff', 'is_superuser',
-            'date_joined', 'last_login',
-        ]
-        for field in account_fields:
-            attr = getattr(account, field, '')
-            if attr:
-                yield (field.replace('_', ' ').title(), attr)
-
-        auth_groups = account.groups.all()
-        html_list = get_html_list(auth_groups)
-        yield ('Auth Groups', html_list or None)
-
-        user_permissions = account.user_permissions.all()
-        html_list = get_html_list(user_permissions)
-        yield ('User Permissions', html_list or None)
-
-        # Polls fields
-        account_name = User._meta.get_field('account')
-        account_name = account_name.verbose_name.title()
-        group_name = User._meta.get_field('group')
-        group_name = group_name.verbose_name.title()
-        for item in super().get_object_data():
-            if item[0] == account_name:
-                continue
-            elif item[0] == group_name:
-                yield ('Polls Group', item[1])
-            else:
-                yield item
-
-        # M2M field
-        followed_users = user.followed_users.order_by('account__username')
-        if len(followed_users):
-            html_list = get_html_list(followed_users)
-            yield ('Followed Users', html_list)
-
-        # Reverse relation
-        followed_question = user.question_follows.order_by('question_text')
-        html_list = get_html_list(followed_question)
-        yield ('Followed Question', html_list or 'None')
-
-        # Relational data
-        question_rel = user.questionfollower_set
-        notify_time = question_rel.filter(notify_time__isnull=False)
-        notify_time = notify_time.values_list('notify_time', flat=True)
-        html_list = get_html_list(notify_time)
-        yield ('Question Notify Times', html_list or 'None')
-
-
-class UserUpdateView(UpdateModelView):
-    def get_form_class(self):
-        return UserForm
-
-
-class UserFilterForm(forms.Form):
-    layout = Layout('search',
-                    'group',
-                    'account__username',
-                    'min_follower_count')
-
-
-class UserFilter(SearchAndFilterSet):
-    search_fields = ['account__username', 'group']
-
-    account__username = CharFilter(lookup_expr='icontains',
-                                   label='Name contains')
-
-    min_follower_count = NumberFilter(field_name='user_followed',
-                                      method='filter_count_gte',
-                                      label='Minimum follower count')
-
-    def filter_count_gte(self, queryset, name, value):
-        qs = queryset.annotate(name_count=Count(name))
-        return qs.filter(name_count__gte=value)
-
-    class Meta:
-        model = User
-        fields = {'group': ['exact']}
-        form = UserFilterForm
-
-
-class UserListView(ListModelView, ListFilterView):
-    list_display = ['name', 'group', 'followers_list']
-    filterset_class = UserFilter
-
-
-class UserViewSet(ModelViewSet):
-    model = User
-    detail_view_class = UserDetailView
-    update_view_class = UserUpdateView
-    list_view_class = UserListView
+from ...utils import (FieldDataMixin, FormSetForm, GetParamAsFormDataMixin,
+                      ListFilterView, NestedModelFormField, SearchAndFilterSet)
 
 
 class AttachmentsForm(FormSetForm):
@@ -438,24 +272,3 @@ class QuestionViewSet(ModelViewSet):
     create_view_class = QuestionCreateView
     update_view_class = QuestionUpdateView
     list_view_class = QuestionListView
-
-
-class VoteViewSet(ModelViewSet):
-    model = Vote
-    list_display = ['timestamp', 'question', 'choice_text', 'is_custom']
-
-    def has_add_permission(self, request, obj=None):
-        return False
-
-    def has_change_permission(self, request, obj=None):
-        return False
-
-
-class PasswordChangeView(generic.View):
-    def get(self, request, *args, **kwargs):
-        return HttpResponse('Form for changing password')
-
-
-class PasswordChangeDoneView(generic.View):
-    def get(self, request, *args, **kwargs):
-        return HttpResponse('Password changed!')
