+++ django_material_demo/cms/polls/views/vote.py
@@ -1,0 +1,13 @@
+from material.frontend.views import ModelViewSet
+from polls.models import Vote
+
+
+class VoteViewSet(ModelViewSet):
+    model = Vote
+    list_display = ['timestamp', 'question', 'choice_text', 'is_custom']
+
+    def has_add_permission(self, request, obj=None):
+        return False
+
+    def has_change_permission(self, request, obj=None):
+        return False
